split TradeAccount into common and broker models
