+++ alpaca/common/models/accounts.py
@@ -10,6 +10,7 @@ from ..enums import (
     OrderStatus,
     OrderSide,
     TradeActivityType,
+    AccountStatus,
 )
 
 
@@ -97,3 +98,80 @@ class TradeActivity(BaseActivity):
     order_id: UUID
     cum_qty: float
     order_status: OrderStatus
+
+
+class TradeAccount(BaseModel):
+    """
+    Represents trading account information for an Account.
+
+    Attributes:
+        id (UUID): The account ID
+        account_number (str): The account number
+        status (AccountStatus): The current status of the account
+        crypto_status (Optional[AccountStatus]): The status of the account in regards to trading crypto. Only present if
+          crypto trading is enabled for your brokerage account.
+        currency (Optional[str]): Currently will always be the value "USD".
+        buying_power (str): Current available cash buying power. If multiplier = 2 then
+          buying_power = max(equity-initial_margin(0) * 2). If multiplier = 1 then buying_power = cash.
+        regt_buying_power (str): User’s buying power under Regulation T
+          (excess equity - (equity - margin value) - * margin multiplier)
+        daytrading_buying_power (str): The buying power for day trades for the account
+        non_marginable_buying_power (str): The non marginable buying power for the account
+        cash (str): Cash balance in the account
+        accrued_fees (str): Fees accrued in this account
+        pending_transfer_out (Optional[str]): Cash pending transfer out of this account
+        pending_transfer_in (Optional[str]): Cash pending transfer into this account
+        portfolio_value (str): Total value of cash + holding positions.
+          (This field is deprecated. It is equivalent to the equity field.)
+        pattern_day_trader (bool): Whether the account is flagged as pattern day trader or not.
+        trading_blocked (bool): If true, the account is not allowed to place orders.
+        transfers_blocked (bool): If true, the account is not allowed to request money transfers.
+        account_blocked (bool): If true, the account activity by user is prohibited.
+        created_at (datetime): Timestamp this account was created at
+        trade_suspended_by_user (bool): If true, the account is not allowed to place orders.
+        multiplier (str): Multiplier value for this account.
+        shorting_enabled (bool): Flag to denote whether or not the account is permitted to short
+        equity (str): This value is cash + long_market_value + short_market_value. This value isn't calculated in the
+          SDK it is computed on the server and we return the raw value here.
+        last_equity (str): Equity as of previous trading day at 16:00:00 ET
+        long_market_value (str): Real-time MtM value of all long positions held in the account
+        short_market_value (str): Real-time MtM value of all short positions held in the account
+        initial_margin (str): Reg T initial margin requirement
+        maintenance_margin (str): Maintenance margin requirement
+        last_maintenance_margin (str): Maintenance margin requirement on the previous trading day
+        sma (str): Value of Special Memorandum Account (will be used at a later date to provide additional buying_power)
+        daytrade_count (int): The current number of daytrades that have been made in the last 5 trading days
+          (inclusive of today)
+    """
+
+    id: UUID
+    account_number: str
+    status: AccountStatus
+    crypto_status: Optional[AccountStatus]
+    currency: Optional[str]
+    buying_power: str
+    regt_buying_power: str
+    daytrading_buying_power: str
+    non_marginable_buying_power: str
+    cash: str
+    accrued_fees: str
+    pending_transfer_out: Optional[str]
+    pending_transfer_in: Optional[str]
+    portfolio_value: str
+    pattern_day_trader: bool
+    trading_blocked: bool
+    transfers_blocked: bool
+    account_blocked: bool
+    created_at: datetime
+    trade_suspended_by_user: bool
+    multiplier: str
+    shorting_enabled: bool
+    equity: str
+    last_equity: str
+    long_market_value: str
+    short_market_value: str
+    initial_margin: str
+    maintenance_margin: str
+    last_maintenance_margin: str
+    sma: str
+    daytrade_count: int
